// blank/src/models/TourismItem.model.ts
export interface TourismItem{
    tourismName: string;
    tourismType: number;
    tourismProvideInformation: string;
    tourismDescription: string;
    tourismLanguage: string;
    tourismDuration: string;
    tourismPrice: string;
    tourismPromotionalPicture: string;
    tourismPromotionalVideo: string;
    tourismLocation: string;
    tourismAvailableDays: string;
    tourismHostName: string;
}